+++ docs/source/change_log.rst
@@ -4,6 +4,8 @@ Change Log
 ``Master branch``
 ^^^^^^^^^^^^^^^^^
 
+* Fix json-bug in create_coveragestore method (see issue `#86 <https://github.com/gicait/geoserver-rest/issues/86>`)
+
 ``[v2.4.0 - 2023-01-10]``
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 * Fix the import issue, close `#76 <https://github.com/gicait/geoserver-rest/issues/76>`_